First check in of the 2.0.2 release. Main changes: - introduced FF_ESAVER (Energy SAVER) compiler flag (experimental) (thanks to Mehdi Goli ) - ff_relax now uses nanosleep instead of usleep - first full porting to ARM processor (thanks to Mauro Mulatero) - added new interfaces for the ff_farm - added a new construct ff_pipe only for C++11 - fixed some problems with cmake compilation - added ff_send_out_to as a method of the ff_loadbalancer class - added parallel_for.hpp which implements 'OpenMP-like' parallel_for - added cleanup_workers method in the farm skeleton - removed fallback function in the farm emitter - removed many warnings when compiling with the Intel compiler - fixed get_my_id() problem for pipeline

+++ tests/test_torus.cpp
@@ -42,14 +42,14 @@ struct masterStage: public ff_node {
     int svc_init() { eossent=false; return 0;}
     void * svc(void * task) {
         if (task==NULL) {
-            for(int i=0;i<bigbatchSize;++i)
+            for(long i=0;i<bigbatchSize;++i)
                 ff_send_out((void*)(i+1));
             if (numBatch>0) --numBatch;
             return GO_ON;
         }
         
         if (numBatch) {
-            for(int i=0;i<smallbatchSize;++i)
+            for(long i=0;i<smallbatchSize;++i)
                 ff_send_out((void*)(i+1));
             --numBatch;
         } else if (!eossent) {
